refactor(tests): Categories COMPLETE - 8 files + 2 helpers, 152 tests passing
- Service: 5 files (crud, hierarchy, updates, utilities, edge-cases) - 105 tests
- Controller: 3 files (crud, tree, updates) - 47 tests
- Helpers: categories.test-helpers.ts + categories-controller.test-helpers.ts
- Original files deleted: categories.service.spec.ts (1,364), categories.controller.spec.ts (832)
- Total: 2,196 lines → 8 focused files (~220 lines avg)
